Improve usage of TagId field
Add-AddressBookEntry: No longer assumes you want to add the user to the
frequent list and uses the first letter of the Name field for Title1.
Now you must explicitly specify which fields use.

Update-AddressBookEntry: Previously never had an option to update the
TagId. It is now using the same implementation as Add-AddressBookEntry.

Added another exported function to ease the usage in calculating the
TagId field based on a letter.