Add implementation of SkipCertificateCheck
This parameter is added to Invoke-WebRequest in PowerShell 6.0, which
also does not respect changes to
[System.Net.ServicePointManager]::ServerCertificateValidationCallback,
so the only way to make this module work in those versions is to use the
-SkipCertificateCheck. I added the extra parameter to all commands and
an implementation for versions that do not have that parameter naturally
on Invoke-WebRequest. This implementation just manually sets the
ServerCertificateValidationCallback to a scriptblock that always returns
true, calls Invoke-WebRequest like normal, and then resets the callback
to whatever it was set to before.