Fix Get-TagId $Frequent value
If the frequent property is not desired, it should be omitted from the
tagId field. Therefore, the previous behavior of casting a bool to a
byte (resulting in a 0 if -Frequent:$false was passed in) was
unnecessary. It didn't seem to impact anything, as it looked like the
scanner just ignored that 0 value, but the simpler approach is to just
see if that parameter is $true or not.