Fix bulk add when objects use different parameters
This is the same issue that affected the bulk updates as well. Each
invocation of the process block would reset $PSBoundParameters to their
default values if they are not included, but won't remove them from the
object entirely. This meant that the .ContainsKey() check would indicate
that the value was passed, even though that was a leftover from a
previous pipeline value and was really empty (or whatever the default
value for that data type is).