Add comment to Add-AddressBookEntry describing the loop
The function operates in batches of 50, because scanners that have more
than that in their address book will return an error if more than 50
entries are requested at once. This is unique to the getObjectsProps
endpoint (which this function uses), as the internal
Search-AddressBookEntry function (which uses the searchObjects endpoint)
will successfully return but truncate the results to 50.