Switch -Frequent back to a switch
This only reason -IsSender and -IsDestination actually need to be a
[nullable[bool]] with the current implementation is that it needs to
know whether they were even applied. Both Add-AddressBookEntry &
Update-AddressBookEntry need to know this so that they can apply the
default of $true to $IsDestination only if not supplied, and
Update-AddressBookEntry needs to know this to know whether to even
update isSender or isDestination. They cannot use
$PSBoundParameters.ContainsKey() for this, because if values were
supplied in the pipeline, once one pipeline entry supplies a value for
one of those parameters, each subsequent process block execution would
return true from
$PSBoundParameters.ContainsKey('IsSender/IsDestination'), but the
property itself would use the default value for the data type, which for
a switch is $false. This means that Update-AddressBookEntry might try
to set isDestination to false when it should not even try to update that
parameter, which could be dangerous as it would effectively disable
those address book entries that follow..

-Frequent does not have this limitation, as using the default switch
value of $false is fine. Even if it used the same [nullable[bool]] type
the other two use, it would reset the value of tagId to not include the
frequent flag ("1") if -Frequent is not provided but other tags are
(although it also cannot use $PSBoundParameters.ContainsKey() for the
same reason).