Wait to add new entry until the end of the block
In Add-AddressBookEntry, if an error occurs while setting properties,
the idea is that once better error handling is setup, the rest of the
function would execute and the other values of the pipeline could be
added. In those cases, we do not want a partial entry to be added to the
XML body, so move the AppendChild to the end of the block, at the end of
where the try {} block would be, where it will only execute if the node
has been completely configured.