Add better error handling
If an error occurs while connecting to the scanner, no cleanup is
necessary. If an error occurs after connection but in the begin block
(probably due to a bad template file), disconnect and terminate the
command. If an error occurs in the process block of an editing function
(Update, Add, or Remove), write a non-terminating error to the host
(using Write-Error), do nothing else, and skip to the next item in the
pipeline. Other values in the pipeline will still be processed. This
will not prevent the disconnection that occurs in the end block. Errors
that occur in the end block will not prevent the disconnection either,
even if they are terminating errors.