fix(add): default -Frequent to $true
See #2. 6916dd2 threw an error message if no -Frequent, -Title1,
-Title2, or -Title3 was provided. This was done so that the address book
entry was findable; the web UI won't let you submit an entry without at
least one of them, because without them the entry will not be visible in
any of the tabs on the scanner (but interestingly, the API will let you
create an entry without them).

However, this is counter-intuitive, because the web UI defaults Frequent
to On. Some alternatives to the current setup is to make Frequent true
if no other tag values are provided, or use Get-Title1Tag (again, only
if none of the other tag values are provided) on the first letter of
Name (which makes sense but makes assumptions about how users want their
address books organized). I've leaned towards making Frequent a boolean
rather than a switch parameter, and making it default to true, like the
web UI. Making it false would then require explicitly providing false to
it ("-Frequent:$false" or "-Frequent $false").