test(add): add test to check an empty request body
If there is a failure in the process block (most likely because there is
no tagId, if -Frequent is set to false and no alternative, either
-Title1, -Title2, or -Title3, is provided), then that entry will not be
added to the scanner. If there are no entries to submit (meaning the
<propListList /> element of the request body is empty) but the request
is sent anyways, the server response with a COMMON_BAD_PARAMETER error.

This test is just to ensure that nothing is sent if every value of the
pipeline errors and there is nothing to add.