test: import custom assertions to tests
The Remove-Module invocations in the AfterAll {} block are to clean up
the imported assertions in Pester. Without removing the Pester module
from the session, the next test run will try to import the custom
assertions module again, which will fail with an error message stating
that the given assertions have already been added.