test: refactor imports
Dot-sourcing a file allows the setup file to also import the
RicohAddressBook module itself. It can also remove an existing module,
if it was imported normally in the same session, so that the import of
the current file does not cause an error, and provide a cleanup function
to remove all imported modules.